added render_per_pixel.c file and moved functions to it
